refactor: request 관련 parser 객체 제거. 각 domain에서 문자열 정적 팩토리 메서드로 이동
