+++ src/main/java/webserver/http/config/WebServerConfig.java
@@ -5,13 +5,7 @@ import webserver.http.domain.controller.RequestProcessor;
 import webserver.http.domain.controller.RootController;
 import webserver.http.domain.controller.StaticResourceController;
 import webserver.http.view.CookiesParser;
-import webserver.http.view.HeadersParser;
-import webserver.http.view.KeyValuePairParser;
-import webserver.http.view.ProtocolParser;
-import webserver.http.view.request.ParametersParser;
-import webserver.http.view.request.RequestLineParser;
 import webserver.http.view.request.RequestReader;
-import webserver.http.view.request.URIParser;
 import webserver.http.view.response.ResponseWriter;
 
 import java.util.ArrayList;
@@ -19,15 +13,9 @@ import java.util.List;
 
 public class WebServerConfig {
     public static RequestReader requestReader() {
-        KeyValuePairParser keyValuePairParser = new KeyValuePairParser();
-        ParametersParser parametersParser = new ParametersParser(keyValuePairParser);
-        URIParser uriParser = new URIParser(keyValuePairParser, parametersParser);
-        CookiesParser cookiesParser = new CookiesParser(keyValuePairParser);
+        CookiesParser cookiesParser = new CookiesParser();
 
         return new RequestReader(
-                new RequestLineParser(uriParser, new ProtocolParser()),
-                new HeadersParser(keyValuePairParser),
-                parametersParser,
                 cookiesParser
         );
     }
