+++ src/main/java/webserver/http/domain/request/RequestLine.java
@@ -1,10 +1,15 @@
 package webserver.http.domain.request;
 
 import webserver.http.domain.Protocol;
+import webserver.http.domain.exception.BadRequestException;
 
 import java.nio.charset.Charset;
 
 public class RequestLine {
+
+    private static final String REQUEST_LINE_DELIMITER_REGEX = " ";
+
+    private static final int REQUEST_LINE_SPLIT_SIZE = 3;
     private final Method method;
     private final URI uri;
     private final Protocol protocol;
@@ -33,6 +38,23 @@ public class RequestLine {
 
     public void decodeCharacter(Charset charset) {
         uri.decodeCharacter(charset);
+    }
+
+    public static RequestLine from(String message) {
+        String[] splitMessage = message.split(REQUEST_LINE_DELIMITER_REGEX);
+        if (splitMessage.length != REQUEST_LINE_SPLIT_SIZE) {
+            throw new BadRequestException(String.format("'[Method] [URI] [Protocol]' 형식의 requestLine 메시지가 아닙니다. {message=%s}", message));
+        }
+
+        String method = splitMessage[0];
+        String uri = splitMessage[1];
+        String protocol = splitMessage[2];
+
+        return new RequestLine(
+                Method.from(method),
+                URI.from(uri),
+                Protocol.from(protocol)
+        );
     }
 
     @Override
