+++ src/main/java/webserver/http/domain/request/URI.java
@@ -1,9 +1,13 @@
 package webserver.http.domain.request;
 
+import webserver.http.domain.KeyValuePair;
+
 import java.nio.charset.Charset;
 import java.util.HashMap;
 
 public class URI {
+    private static final String QUERY_PARAMETER_DELIMITER = "?";
+
     private final String path;
     private final Parameters queryParameters;
 
@@ -30,6 +34,13 @@ public class URI {
 
     public void decodeCharacter(Charset charset) {
         queryParameters.decodeCharacter(charset);
+    }
+
+    public static URI from(String message) {
+        KeyValuePair pathQueryParamPair = KeyValuePair.from(message, QUERY_PARAMETER_DELIMITER, false);
+        String path = pathQueryParamPair.getKey();
+        String queryParameters = pathQueryParamPair.getValue();
+        return new URI(path, Parameters.from(queryParameters));
     }
 
     @Override
