+++ src/main/java/webserver/http/view/request/RequestReader.java
@@ -1,14 +1,13 @@
 package webserver.http.view.request;
 
-import webserver.utils.IOUtils;
 import webserver.http.domain.Cookies;
+import webserver.http.domain.Headers;
+import webserver.http.domain.exception.NullRequestException;
 import webserver.http.domain.request.Parameters;
 import webserver.http.domain.request.Request;
 import webserver.http.domain.request.RequestLine;
 import webserver.http.view.CookiesParser;
-import webserver.http.domain.Headers;
-import webserver.http.domain.exception.NullRequestException;
-import webserver.http.view.HeadersParser;
+import webserver.utils.IOUtils;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -16,15 +15,9 @@ import java.util.List;
 import java.util.Objects;
 
 public class RequestReader {
-    private final RequestLineParser requestLineParser;
-    private final HeadersParser headersParser;
-    private final ParametersParser requestBodyParser;
     private final CookiesParser cookiesParser;
 
-    public RequestReader(RequestLineParser requestLineParser, HeadersParser headersParser, ParametersParser requestBodyParser, CookiesParser cookiesParser) {
-        this.requestLineParser = requestLineParser;
-        this.headersParser = headersParser;
-        this.requestBodyParser = requestBodyParser;
+    public RequestReader(CookiesParser cookiesParser) {
         this.cookiesParser = cookiesParser;
     }
 
@@ -36,13 +29,13 @@ public class RequestReader {
         List<String> headerMessages = IOUtils.readWhileEmptyLine(bufferedReader);
         headerMessages.forEach(System.out::println);
 
-        RequestLine requestLine = requestLineParser.parse(requestLineMessage);
-        Headers requestHeaders = headersParser.parse(headerMessages);
+        RequestLine requestLine = RequestLine.from(requestLineMessage);
+        Headers requestHeaders = Headers.from(headerMessages);
         Request request = parseRequestExcludeBody(requestLine, requestHeaders);
 
         if (request.hasContents() && request.hasContentType("application/x-www-form-urlencoded")) {
             String requestBody = IOUtils.readData(bufferedReader, request.getContentLength());
-            Parameters parameters = requestBodyParser.parse(requestBody);
+            Parameters parameters = Parameters.from(requestBody);
             request.addParameters(parameters);
         }
