+++ src/test/java/webserver/http/domain/request/RequestLineTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import webserver.http.domain.Protocol;
+import webserver.http.domain.exception.BadRequestException;
 import webserver.http.domain.request.Method;
 import webserver.http.domain.request.Parameters;
 import webserver.http.domain.request.RequestLine;
@@ -17,7 +18,9 @@ import java.util.Map;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
+import static webserver.http.domain.request.fixture.URIFixture.fixtureWithQueryParameters;
 
 class RequestLineTest {
 
@@ -127,5 +130,29 @@ class RequestLineTest {
                 new URI("/path", addedParameters),
                 new Protocol("HTTP", "1.1")
         );
+    }
+
+    @DisplayName("문자열이 white space 값을 기준으로 3개의 문자열로 구분이 되지 않는 경우 예외발생")
+    @Test
+    void parse_fail() {
+        String invalidRequestLineMessage = "GET /uri?name=jordy PROTO/1,1 OTHER-DATA";
+        assertThatThrownBy(() -> RequestLine.from(invalidRequestLineMessage))
+                .isInstanceOf(BadRequestException.class)
+                .hasMessage("'[Method] [URI] [Protocol]' 형식의 requestLine 메시지가 아닙니다. {message=" + invalidRequestLineMessage + "}");
+    }
+
+    @DisplayName("문자열이 white space 값을 기준으로 3개의 문자열로 구분되는 경우 RequestLine 객체를 생성한다.")
+    @Test
+    void parse() {
+        String requestLineMessage = "GET /uri?name=jordy HTTP/1.1";
+
+        RequestLine actual = RequestLine.from(requestLineMessage);
+
+        assertThat(actual).usingRecursiveComparison()
+                .isEqualTo(new RequestLine(
+                        Method.GET,
+                        fixtureWithQueryParameters("/uri", "name", "jordy"),
+                        new Protocol("HTTP", "1.1")
+                ));
     }
 }
