+++ src/test/java/webserver/http/domain/request/URITest.java
@@ -6,8 +6,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import webserver.http.domain.request.Parameters;
-import webserver.http.domain.request.URI;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -15,8 +13,27 @@ import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
+import static webserver.http.domain.request.fixture.URIFixture.fixtureWithQueryParameters;
 
 class URITest {
+
+    @DisplayName("queryParameter가 포함되지 않는 메시지인 경우, path와 빈 QueryParameters를 갖는 URI 객체를 생성한다.")
+    @Test
+    void parse_empty_query_parameter() {
+        String message = "/hello";
+        URI actual = URI.from(message);
+        assertThat(actual).usingRecursiveComparison()
+                .isEqualTo(new URI("/hello"));
+    }
+
+    @DisplayName("queryParameter가 포함된 메시지인 경우, path와 빈 QueryParameters를 갖는 URI 객체를 생성한다.")
+    @Test
+    void parse() {
+        String message = "/hello?name=jordy";
+        URI actual = URI.from(message);
+        assertThat(actual).usingRecursiveComparison()
+                .isEqualTo(fixtureWithQueryParameters("/hello", "name", "jordy"));
+    }
 
     @DisplayName("Parameter에 저장된 key 에 해당하는 value를 가져온다.")
     @ParameterizedTest
